+++ babel.config.js
@@ -20,7 +20,7 @@ module.exports = {
   ],
   env: {
     test: {
-      plugins: ["transform-es2015-modules-commonjs"]
+      plugins: ["@babel/plugin-transform-modules-commonjs"]
     }
   }
 };