+++ packages/reakit/src/Grid/index.ts
@@ -1,15 +1,9 @@
 import Grid from "./Grid";
 import GridItem from "./GridItem";
 
-export interface GridComponents {
-  Item: typeof GridItem;
-}
-
-const G = Grid as typeof Grid & GridComponents;
-
-G.Item = GridItem;
-
 export * from "./GridItem";
 export * from "./Grid";
 
-export default G;
+export default Object.assign(Grid, {
+  Item: GridItem
+});