+++ packages/reakit/src/Group/index.ts
@@ -1,0 +1,9 @@
+import Group from "./Group";
+import GroupItem from "./GroupItem";
+
+export * from "./Group";
+export * from "./GroupItem";
+
+export default Object.assign(Group, {
+  Item: GroupItem
+});