+++ packages/reakit/src/Hidden/__tests__/HiddenContainer-test.ts
@@ -1,18 +1,29 @@
-import { initialState, toggle, show, hide } from "../HiddenContainer";
+import HiddenContainer from "../HiddenContainer";
+import callMeMaybe from "../../_utils/callMeMaybe";
+
+const { initialState, actions: a } = HiddenContainer;
 
 test("initialState", () => {
   expect(initialState).toEqual({ visible: false });
 });
 
 test("toggle", () => {
-  expect(toggle()({ visible: false })).toEqual({ visible: true });
-  expect(toggle()({ visible: true })).toEqual({ visible: false });
+  expect(callMeMaybe(a.toggle(), { visible: false })).toEqual({
+    visible: true
+  });
+  expect(callMeMaybe(a.toggle(), { visible: true })).toEqual({
+    visible: false
+  });
 });
 
 test("show", () => {
-  expect(show()()).toEqual({ visible: true });
+  expect(callMeMaybe(a.show(), initialState)).toEqual({
+    visible: true
+  });
 });
 
 test("hide", () => {
-  expect(hide()()).toEqual({ visible: false });
+  expect(callMeMaybe(a.hide(), initialState)).toEqual({
+    visible: false
+  });
 });