+++ packages/reakit/src/Step/StepNext.tsx
@@ -1,12 +1,20 @@
-import React from "react";
-import PropTypes from "prop-types";
+import * as React from "react";
+import * as PropTypes from "prop-types";
 import { theme } from "styled-tools";
 import callAll from "../_utils/callAll";
 import styled from "../styled";
 import as from "../as";
 import Box from "../Box";
+import { StepContainerSelectors, StepContainerActions } from "./StepContainer";
 
-const Component = ({ onClick, ...props }) => (
+export interface StepNextProps {
+  next: StepContainerActions["next"];
+  hasNext?: StepContainerSelectors["hasNext"];
+  loop?: boolean;
+  onClick?: () => any;
+}
+
+const Component = ({ onClick, ...props }: StepNextProps) => (
   <Box
     onClick={callAll(props.next, onClick)}
     disabled={!props.loop && props.hasNext && !props.hasNext()}
@@ -18,6 +26,7 @@ const StepNext = styled(Component)`
   ${theme("StepNext")};
 `;
 
+// @ts-ignore
 StepNext.propTypes = {
   next: PropTypes.func.isRequired,
   hasNext: PropTypes.func,