+++ packages/reakit/src/Toolbar/ToolbarFocusable.tsx
@@ -1,22 +1,38 @@
-import React from "react";
+import * as React from "react";
 import { findDOMNode } from "react-dom";
-import PropTypes from "prop-types";
+import * as PropTypes from "prop-types";
 import { theme } from "styled-tools";
 import hoistNonReactStatics from "hoist-non-react-statics";
 import callAll from "../_utils/callAll";
+import getSelector from "../_utils/getSelector";
 import styled from "../styled";
 import as from "../as";
 import Box from "../Box";
 import Toolbar from "./Toolbar";
 
-class Component extends React.Component {
-  static propTypes = {
-    tabIndex: PropTypes.number,
-    onFocus: PropTypes.func,
-    disabled: PropTypes.bool
-  };
+type Focusable = (Element | Text) & {
+  focus(): void;
+};
 
-  constructor(props) {
+export interface ToolbarFocusableProps {
+  disabled?: boolean;
+  tabIndex?: number;
+  onFocus?: () => void;
+}
+
+export interface ToolbarFocusableState {
+  tabIndex: number;
+}
+
+class ToolbarFocusableComponent extends React.Component<
+  ToolbarFocusableProps,
+  ToolbarFocusableState
+> {
+  private element: any;
+
+  private toolbar: any;
+
+  constructor(props: ToolbarFocusableProps) {
     super(props);
     this.state = {
       tabIndex: this.getInitialTabIndex()
@@ -33,7 +49,7 @@ class Component extends React.Component {
     }
   }
 
-  componentDidUpdate(prevProps) {
+  componentDidUpdate(prevProps: ToolbarFocusableProps) {
     const { disabled } = this.props;
 
     if (prevProps.disabled !== disabled) {
@@ -63,21 +79,19 @@ class Component extends React.Component {
 
   getToolbar = () => {
     if (typeof this.toolbar === "undefined") {
-      this.toolbar = this.getElement().closest(`.${Toolbar.styledComponentId}`);
+      this.toolbar = this.getElement().closest(getSelector(Toolbar));
     }
     return this.toolbar;
   };
 
-  getFocusables = () => {
-    if (!this.getToolbar()) return [];
-    return this.getToolbar().querySelectorAll(
-      `.${ToolbarFocusable.styledComponentId}`
-    );
+  getFocusables = (): NodeListOf<Focusable> => {
+    if (!this.getToolbar()) return new NodeList() as NodeListOf<Focusable>;
+    return this.getToolbar().querySelectorAll(getSelector(ToolbarFocusable));
   };
 
-  getCurrentIndex = focusables => {
+  getCurrentIndex = (focusables: NodeListOf<Focusable>) => {
     let currentIndex = -1;
-    focusables.forEach((item, i) => {
+    focusables.forEach((item: Element | Text | null, i: number) => {
       if (item === this.getElement()) {
         currentIndex = i;
       }
@@ -85,12 +99,18 @@ class Component extends React.Component {
     return currentIndex;
   };
 
-  getNextFocusable = (focusables, currentIndex) => {
+  getNextFocusable = (
+    focusables: NodeListOf<Focusable>,
+    currentIndex: number
+  ) => {
     const index = currentIndex + 1;
     return focusables.item(index) || focusables.item(0);
   };
 
-  getPreviousFocusable = (focusables, currentIndex) => {
+  getPreviousFocusable = (
+    focusables: NodeListOf<Focusable>,
+    currentIndex: number
+  ) => {
     const index = currentIndex ? currentIndex - 1 : focusables.length - 1;
     return focusables.item(index);
   };
@@ -110,7 +130,7 @@ class Component extends React.Component {
     this.getElement().removeEventListener("keydown", this.handleKeyDown);
   };
 
-  handleKeyDown = e => {
+  handleKeyDown = (e: KeyboardEvent) => {
     const isVertical = this.toolbarIsVertical();
     const nextKey = isVertical ? "ArrowDown" : "ArrowRight";
     const previousKey = isVertical ? "ArrowUp" : "ArrowLeft";
@@ -146,10 +166,17 @@ class Component extends React.Component {
   }
 }
 
-hoistNonReactStatics(Component, Box);
+hoistNonReactStatics(ToolbarFocusableComponent, Box);
 
-const ToolbarFocusable = styled(Component)`
+const ToolbarFocusable = styled(ToolbarFocusableComponent)`
   ${theme("ToolbarFocusable")};
 `;
+
+// @ts-ignore
+ToolbarFocusable.propTypes = {
+  tabIndex: PropTypes.number,
+  onFocus: PropTypes.func,
+  disabled: PropTypes.bool
+};
 
 export default as("div")(ToolbarFocusable);