+++ packages/reakit/src/_utils/createElementRef.js
@@ -1,9 +1,0 @@
-/* eslint-disable no-param-reassign */
-const createElementRef = (scope, property) => element => {
-  scope[property] = element;
-  if (scope.props && scope.props.elementRef) {
-    scope.props.elementRef(element);
-  }
-};
-
-export default createElementRef;