+++ packages/reakit/src/_utils/getSelector.ts
@@ -1,0 +1,8 @@
+function getSelector(object: any): string {
+  if (object != null && object.styledComponentId) {
+    return `.${object.styledComponentId}`;
+  }
+  return "";
+}
+
+export default getSelector;