+++ packages/reakit/src/styled.ts
@@ -21,4 +21,5 @@ export {
   ServerStyleSheet,
   StyleSheetManager
 };
+
 export default styled;